server/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
psycopg2==2.9.9
PyJWT==2.9.0
PyYAML==6.0.1
SQLAlchemy==2.0.32